==> 1-error-detection-dart-basiscript.ipynb <==
import pandas as pd
import os
import glob
import re

# ADAPT
import alignment_adapt.run_adapt as run_adapt
import alignment_adapt.cgn2_adapt_map as cgn2_adapt_map
import alignment_adapt.deduce_pcus_orig_phon as deduce_pcus

#ADAGT
import alignment_adagt.adagt as adagt
import alignment_adagt.deduce_pcus_orig_graph as deduce_pcus_graph


# INPUTS - Language Resources

Phonetic Lexicons

lrecDataDir = '../../astla-data/lrec-data'
phoneticLexiconFileDART = os.path.join(lrecDataDir, '2-lexicon-dart-preposttest-checked.tsv')
phoneticLexiconFileBS= os.path.join(lrecDataDir, '2-lexicon-basiscript-g4p1-checked.tsv')
phoneticLexiconFileDART_BS = os.path.join(lrecDataDir, '2-lexicon.tsv')

with open(phoneticLexiconFileDART, 'r') as f:
    lexiconDART = [line.replace('\n', '').split('\t') for line in f.readlines()]

with open(phoneticLexiconFileBS, 'r') as f:
    lexiconBS = [line.replace('\n', '').split('\t') for line in f.readlines()]

lexiconDictDART= pd.DataFrame(lexiconDART).set_index(0)
lexiconDictBS = pd.DataFrame(lexiconBS).set_index(0)
lexicon_dict = pd.concat([lexiconDictDART,lexiconDictBS]).drop_duplicates().sort_index()

lexicon_dict.to_csv(phoneticLexiconFileDART_BS)
lexicon_dict.loc['bakken', 1]

Aligned Phonetic lexicons

lrecDataDir = '../../astla-data/lrec-data'
alignedPhoneticLexiconFileDART = os.path.join(lrecDataDir, '3-aligned-lexicon-dart-preposttest.csv')
alignedPhoneticLexiconFileBS= os.path.join(lrecDataDir, '3-aligned-lexicon-basiscript-g4p1.csv')
alignedPhoneticLexiconFileDART_BS= os.path.join(lrecDataDir, '3-aligned-lexicon.csv')

pcuLexiconDART = pd.read_csv(alignedPhoneticLexiconFileDART)
pcuLexiconBS = pd.read_csv(alignedPhoneticLexiconFileBS)

pcu_lexicon = pd.concat([pcuLexiconDART,pcuLexiconBS]).drop_duplicates().sort_index()

# Convert strings to list of strings
pcu_lexicon['graphemes_align'] = pcu_lexicon['graphemes_align'].apply(lambda x: eval(x))
pcu_lexicon['phonemes_align'] = pcu_lexicon['phonemes_align'].apply(lambda x: eval(x))

pcu_lexicon.to_csv(alignedPhoneticLexiconFileDART_BS)
pcu_lexicon

Sound Pure PCUs

klankzuiver_dict = {
    "a": ["aa"],
    "A": ["a"],
    "b": ["b"],
    "d": ["d"],
    "e": ["ee"], 
    "E": ["e"],
    "EI": ["ei", "ij"],
    "f": ["f"],
    "G": ["g", "ch"],
    "h": ["h"],
    "i": ["ie"],
    "I": ["i"],
    "j": ["j"],
    "k": ["k"],
    "l": ["l"],
    "m": ["m"],
    "n": ["n"],
    "N": ["ng"],
    "o": ["oo"],
    "EU": ["eu"],
    "U": ["u"],
    "UI": ["ui"],
    "O": ["o"],
    "p": ["p"],
    "r": ["r"],
    "s": ["s"],
    "t": ["t"],
    "u": ["oe"],
    "v": ["v"],
    "w": ["w"], 
    "AU": ["au", "ou"],
    "x": ["g", "ch"],
    "y": ["uu"],
    "U": ["u"],
    "z": ["z"],
}

# Getter & Setter Functions

def getPhoneticTranscription(prompt):
    try:
        return lexicon_dict.loc[prompt, 1]
    except:
        return '<unk>'
    
getPhoneticTranscription('blauw')

# Functions
def getPCUS(pcu_lexicon, prompt):
    try:
        selected_row = pcu_lexicon[pcu_lexicon['graphemes'] == prompt].reset_index()
        pcus_aligned = selected_row.loc[0,'graphemes_align']
        phonemes_aligned = selected_row.loc[0,'phonemes_align']
        return pcus_aligned, phonemes_aligned
    
    except:
        print('NOT IN LEXION:', prompt)
        return [],[]
    
getPCUS(pcu_lexicon, 'web') # web, worstje

"""
Input: 
pcus_aligned: ['w', 'e', 'b']
phones_aligned: ['w', 'E', 'p']

Output:
For each PCU-phoneme pair whether it is sound pure (klankzuiver) or not: [True, True, False]
"""
def isKlankzuiver(graphemes_aligned, phonemes_aligned):
    return [phone in klankzuiver_dict.keys() and graphemes_aligned[idx] in klankzuiver_dict[phone] for idx, phone in enumerate(phonemes_aligned)]

# Label each target word with "klankzuiver": False or True
def wordKlankzuiver(row):
    return row['klankzuiver_pcus_total'] == len(row['target_phones']) and row['klankzuiver_pcus_total']>0

isKlankzuiver(getPCUS(pcu_lexicon, 'web')[0], getPCUS(pcu_lexicon, 'web')[1])

# Read original data

DART read word lists

### Fill DART DF List with different dart preposttest subsets with annotations
dart_df_list = []

### DART: Stephanies annotations (SPECOM set)
csv_file_list = glob.glob('../../astla-data/dart-preposttest/specom-data/chunks_attempts_matched_6sep_manAnn_whispert/*.csv')

for csv_file in csv_file_list:

    # Read all annotations
    df = pd.read_csv(csv_file).fillna('<unk>')
    
    # Add phonetic transcription of prompt
    df['prompt_in_phonemes'] = df['prompt'].apply(getPhoneticTranscription)
    df['phonTrans'] = df['phonTrans'].apply(lambda x: re.sub('\s\s+', ' ', x.replace('-', ' ')))
    # print(df['assessmentDescription'])

    # Select relevant data
    df_selection = df[['index', 'prompt', 'prompt_in_phonemes', 'phonTrans', 'assessmentDescription']].rename(columns={'index': 'filename', 'assessmentDescription': 'error_cat'})
    
    dart_df_list.append(df_selection)

print(len(dart_df_list))

### DART Core set
dart_core_df = pd.read_csv('../../astla-data/dart-preposttest/slate-data/10-output/dart-coreset-all-information.csv')
# print(dart_core_df['phon_category'])

def extractPromptFromFilename(filename):
    return filename.split('_')[1]
    
dart_core_df['prompt'] = dart_core_df['filename'].apply(extractPromptFromFilename)
dart_core_df['prompt_in_phonemes_prep']= dart_core_df['prompt_in_phonemes'].apply(lambda x: re.sub('\s\s+', ' ', x.replace('-', ' ')))

dart_df = dart_core_df.loc[:, ['filename', 'prompt', 'prompt_in_phonemes_prep', 'phonetician1', 'phon_category']].rename(columns={'phonetician1': 'phonTrans', 'prompt_in_phonemes_prep': 'prompt_in_phonemes', 'phon_category':'error_cat'})
dart_df.head(2)

dart_df_list.append(dart_df)
print(len(dart_df_list))


# Concatenate both DART datasets
dart_df = pd.concat(dart_df_list).reset_index(drop=True)
dart_df.to_csv('/vol/tensusers5/wharmsen/astla-data/lrec-data/dart-reading-selection.csv')
dart_df

Basiscript dictations

# BasiScript
basiscript_df = pd.read_csv('/vol/tensusers5/wharmsen/spelling-data/aseda/3-classdata/word-dataframe-basiscript-dictee.csv', converters={'target_phones': eval,
                                                           'target_pcus': eval,
                                                           'original_pcus': eval})

def isGroep4p1(pair_id):
    return pair_id.find('.g4.p1') != -1

def isGroep4p2(pair_id):
    return pair_id.find('.g4.p2') != -1

groep4DF = basiscript_df[basiscript_df['pair_id'].apply(isGroep4p1)]
groep4p2DF = basiscript_df[basiscript_df['pair_id'].apply(isGroep4p2)]

groep4DF.head()

# Restore Student IDs
studentIDlist = []
for i in range(int(len(groep4DF)/25)):
    studentID = 's' + str(i)
    studentIDlist += [studentID] * 25

groep4DF.loc[:,'studentID'] = studentIDlist

# Make data selection and add prompts
groep4DF.loc[:, 'target'] = groep4DF.loc[:,'target_aligned'].apply(lambda x: x.replace("*", "") )
groep4DF = groep4DF[['target', 'original', 'studentID']].reset_index(drop=True)

# Alignment 1: Target graph-phon

pcu_lexicon contains for each prompt the alignment between target graphemes and phonemes.
For each prompt in the data set, we look up the corresponding alignment in this pcu_lexicon

DART

target_pcus = []
target_phones = []
klankzuiver = []

for idx, row in dart_df.iterrows():
    prompt = dart_df.loc[idx, 'prompt']
    pcus_aligned, phonemes_aligned = getPCUS(pcu_lexicon, prompt)

    pcus_isKlankzuiver = isKlankzuiver(pcus_aligned, phonemes_aligned)

    target_pcus.append(pcus_aligned)
    target_phones.append(phonemes_aligned)
    klankzuiver.append(pcus_isKlankzuiver)

dart_df['target_phones'] = target_phones
dart_df['target_pcus'] = target_pcus
dart_df['klankzuiver'] = klankzuiver
dart_df['klankzuiver_pcus_total'] = dart_df['klankzuiver'].apply(lambda x: sum(x))
dart_df['word_klankzuiver'] = dart_df.apply(wordKlankzuiver, axis = 1)

dart_df.head(2)

BasiScript

target_pcus = []
target_phones = []
klankzuiver = []

for idx, row in groep4DF.iterrows():
    prompt = groep4DF.loc[idx, 'target']
    pcus_aligned, phonemes_aligned = getPCUS(pcu_lexicon, prompt)

    pcus_isKlankzuiver = isKlankzuiver(pcus_aligned, phonemes_aligned)

    target_pcus.append(pcus_aligned)
    target_phones.append(phonemes_aligned)
    klankzuiver.append(pcus_isKlankzuiver)

groep4DF['target_phones'] = target_phones
groep4DF['target_pcus'] = target_pcus
groep4DF['klankzuiver'] = klankzuiver
groep4DF['klankzuiver_pcus_total'] = groep4DF['klankzuiver'].apply(lambda x: sum(x))
groep4DF['word_klankzuiver'] = groep4DF.apply(wordKlankzuiver, axis = 1)
groep4DF.head(2)

# ADAPT and ADAGT

ADAPT: Align phonetic strings of DART data

# ADAPT alignment

# prompt = "kijken"
# prompt_ref = "k EI k @"
# prompt_hyp = "k EI k @ n"

prompt = 'hagel'
prompt_ref = 'h a x @ l'
prompt_hyp = 'h A N h a x @ r'

# PCU lexicon uses * for ins/del
target_pcus, target_phones = getPCUS(pcu_lexicon, prompt)
target_phones_adapt = [cgn2_adapt_map.cgn2_to_adapt_dict[phone] for phone in target_phones] #o.a. change * to - 
print(target_pcus)
print(target_phones)
print(target_phones_adapt)

# ADAPT uses '-' for ins/del
align_ref, align_hyp, align_ref_adapt, align_hyp_adapt = run_adapt.reverse_align_two_phone_strings(prompt_ref, prompt_hyp)
print(align_ref)
print(align_hyp)
print(align_ref_adapt)
print(align_hyp_adapt)

pcus_target_graph, pcus_target_phon_adapt, pcus_orig_phon_adapt = deduce_pcus.computePCUs(align_ref_adapt, align_hyp_adapt, target_phones_adapt, target_pcus, '-')

# Postprocess
pcus_target_phon = [cgn2_adapt_map.adapt_to_cgn2_dict[phone] for phone in pcus_target_phon_adapt[0]]
pcus_orig_phon = [cgn2_adapt_map.adapt_to_cgn2_dict[phone] for phone in pcus_orig_phon_adapt[0]]
print(pcus_target_graph, pcus_target_phon, pcus_orig_phon)

# Initialize new columns
pcus_target_graph_list = []
pcus_target_phon_list = []
pcus_orig_phon_list = []

for idx, row in dart_df.iterrows():
    # try:
    # Preprocess
    prompt_ref = row['prompt_in_phonemes']
    prompt_hyp = row['phonTrans']

    # The case that the prompt word is not read
    if(prompt_ref == '<unk>' or prompt_hyp == '<unk>'):
        print(prompt_ref, prompt_hyp)
        pcus_target_graph = []
        pcus_target_phon = []
        pcus_orig_phon = []

    else:
        try:
            # ADAPT alignment
            align_ref, align_hyp, align_ref_adapt, align_hyp_adapt = run_adapt.reverse_align_two_phone_strings(prompt_ref, prompt_hyp)

            target_phones = row['target_phones']
            target_phones_adapt = [cgn2_adapt_map.cgn2_to_adapt_dict[phone] for phone in target_phones]

            target_pcus = row['target_pcus']

            # Alignment with original phonemes
            pcus_target_graph, pcus_target_phon_adapt, pcus_orig_phon_adapt = deduce_pcus.computePCUs(align_ref_adapt, align_hyp_adapt, target_phones_adapt, target_pcus, '-')

            # Postprocess
            pcus_target_phon = [cgn2_adapt_map.adapt_to_cgn2_dict[phone] for phone in pcus_target_phon_adapt[0]]
            pcus_orig_phon = [cgn2_adapt_map.adapt_to_cgn2_dict[phone] for phone in pcus_orig_phon_adapt[0]]
        except:
            print(target_pcus)
            print(target_phones)
            print(target_phones_adapt)
            print(align_ref)
            print(align_hyp)
            print(align_ref_adapt)
            print(align_hyp_adapt)
            print(pcus_target_graph, pcus_target_phon_adapt, pcus_orig_phon_adapt)
        
    pcus_target_graph_list.append(pcus_target_graph[0]) if len(pcus_target_graph) > 0 else pcus_target_graph_list.append(pcus_target_graph)
    pcus_target_phon_list.append(pcus_target_phon)
    pcus_orig_phon_list.append(pcus_orig_phon)

dart_df['pcus_target_graph'] = pcus_target_graph_list
dart_df['pcus_target_phon'] = pcus_target_phon_list
dart_df['pcus_orig_phon'] = pcus_orig_phon_list
dart_df.head(3)

pcuSegmFileDART = os.path.join(lrecDataDir, '4-pcu-segmentations-dart-preposttest.csv')
dart_df.to_csv(pcuSegmFileDART, sep=';')

ADAGT: Align grapheme strings of BasiScript data

groep4DF.head(1)

# Initialize new columns
target_aligned_list = []
original_aligned_list = []
pcus_target_phon_list = []
pcus_target_graph_list = []
pcus_orig_graph_list = []

for idx, row in groep4DF.iterrows():

    try:
        # Preprocess
        prompt_ref = row['target']
        prompt_hyp = row['original']

        # ADAGT alignment
        align_ref, align_hyp = adagt.align(prompt_ref, prompt_hyp)

        target_phones = row['target_phones']
        target_pcus = row['target_pcus']

        # Alignment with original phonemes
        pcus_target_phon, pcus_target_graph, pcus_orig_graph = deduce_pcus_graph.computePCUs(align_ref, align_hyp, target_pcus, target_phones, '-')
        
    # Except case is activated in case word is not read (in that case "target_phones" is NaN)
    except:
        align_ref = ''
        align_hyp = ''
        pcus_target_phon = []
        pcus_target_graph = []
        pcus_orig_graph = []

    target_aligned_list.append(align_ref)
    original_aligned_list.append(align_hyp)
    pcus_target_phon_list.append(pcus_target_phon)
    pcus_target_graph_list.append(pcus_target_graph[0]) if len(pcus_target_graph) > 0 else pcus_target_graph_list.append(pcus_target_graph)
    pcus_orig_graph_list.append(pcus_orig_graph)

groep4DF['target_aligned'] = target_aligned_list
groep4DF['original_aligned'] = original_aligned_list
groep4DF['pcus_target_graph'] = pcus_target_graph_list
groep4DF['pcus_target_phon'] = [x[0] if len(x)>0 else [] for x in pcus_target_phon_list ]
groep4DF['pcus_orig_graph'] = [x[0] if len(x)>0 else [] for x in pcus_orig_graph_list ]
groep4DF.head(3)

pcuSegmFileBS = os.path.join(lrecDataDir, '4-pcu-segmentations-basiscript-g4p1.csv')
groep4DF.to_csv(pcuSegmFileBS, sep=';')

Test for adagt and deduce_pcus_graph

# Test function
import alignment_adagt.adagt as adagt
import alignment_adagt.deduce_pcus_orig_graph as deduce_pcus_graph

# Preprocess
prompt_ref = 'monteur'
prompt_hyp = 'montur'

# ADAGT alignment
align_ref, align_hyp = adagt.align(prompt_ref, prompt_hyp)
print(align_ref, align_hyp)

target_phones, target_pcus = getPCUS(pcu_lexicon, align_ref)
print(target_phones, target_pcus)

# Alignment with original phonemes
pcus_target_phon, pcus_target_graph, pcus_orig_graph = deduce_pcus_graph.computePCUs(align_ref, align_hyp, target_pcus, target_phones, '-')
print(pcus_target_phon, pcus_target_graph, pcus_orig_graph)

deduce_pcus_graph.computePCUs('rei-s', 'r-ijs', ['r', 'ei', 's'], ['r', 'EI', 's'], '*')
deduce_pcus_graph.computePCUs('vrouw', 'vraow', ['v', 'r', 'auw'], ['v', 'r', 'AU'], '*')
