==> 2-error-analysis-basiscript.ipynb <==
import pandas as pd
import os

# Read BasiScript data
lrecDataDir = '../../astla-data/lrec-data'
pcuSegmFileBasiScript = os.path.join(lrecDataDir, '4-pcu-segmentations-basiscript-g4p1.csv')
groep4DF = pd.read_csv(pcuSegmFileBasiScript, sep=';', converters = {'target_phones':eval,'target_pcus':eval, 'klankzuiver':eval, 'pcus_target_graph':eval,'pcus_target_phon':eval, 'pcus_orig_graph':eval})
# groep4DF = groep4DF.rename(columns={'pcus_orig_phon':'pcus_orig_graph'})
# groep4DF['pcus_target_phon'] = [x[0] if len(x)>0 else [] for x in groep4DF['pcus_target_phon'] ]
# groep4DF['pcus_orig_graph'] = [x[0] if len(x)>0 else [] for x in groep4DF['pcus_orig_graph'] ]
groep4DF[groep4DF['target'] != groep4DF['original']]

groep4DF[groep4DF['original_aligned'] == '']

### Print Statistics

# 672 words, of which 308 are incorrect
print("OVERALL STATISTICS")
print("Length complete DF: ", len(groep4DF))

print("\nSOUND PURE DART STATISTICS")

# Select sound pure (klankzuiver) words
groep4DF_kz = groep4DF[groep4DF['word_klankzuiver'] == True]

# Print some statistics
print("Length transparent DF: ", len(groep4DF_kz))

correctDF = len(groep4DF_kz[groep4DF_kz['target'] == groep4DF_kz['original']])
print("... of which correct: ", correctDF, ' = ', correctDF/len(groep4DF_kz))

incorrectDF = len(groep4DF_kz[groep4DF_kz['target'] != groep4DF_kz['original']])
print("... of which incorrect: ", incorrectDF, ' = ', incorrectDF/len(groep4DF_kz) )

# Print different prompts
print("Different prompts: ", len(set(groep4DF_kz["target"])))
print("Unique prompts: ", " ".join(sorted(list(set(groep4DF_kz["target"])))))
# dart_df_kz["prompt"].value_counts()

groep4DF_kz['target'].value_counts()

#Switch to phoneme dataframe
pcu_matrix = []
for idx, row in groep4DF_kz.iterrows():
    prompt = row["target"]
    pcus_target_graph = row["pcus_target_graph"]
    pcus_target_phon = row["pcus_target_phon"]
    pcus_orig_graph = row["pcus_orig_graph"]

    if(len(pcus_target_graph) == len(pcus_target_phon) and len(pcus_target_graph) == len(pcus_orig_graph)):
        for idx, item in enumerate(pcus_target_graph):
            pcu_matrix.append([pcus_target_graph[idx], pcus_target_phon[idx], pcus_orig_graph[idx], prompt,  pcus_target_graph[idx] == pcus_orig_graph[idx]])
    else:
        print(pcus_target_graph, pcus_target_phon, pcus_orig_graph)

groep4DF_kz_pcu_df = pd.DataFrame(pcu_matrix, columns = ["pcu", "target_phon", "orig_pcu", "prompt", "correct"])
groep4DF_kz_pcu_df.head(2)

print('total # pcus: ', len(groep4DF_kz_pcu_df))
print('total # correct pcus: ', len(groep4DF_kz_pcu_df[groep4DF_kz_pcu_df['correct'] == True]))
print('total # incorrect pcus: ', len(groep4DF_kz_pcu_df[groep4DF_kz_pcu_df['correct'] == False]))

groep4DF_kz_pcu_df[groep4DF_kz_pcu_df['correct'] == True]

groep4DF_kz_pcu_df[groep4DF_kz_pcu_df['correct'] == False]["pcu"].value_counts()

unique_inc_pcus = list(groep4DF_kz_pcu_df[groep4DF_kz_pcu_df['correct'] == False]["pcu"].value_counts().index)
unique_inc_pcus

outputMatrix = []
for unique_pcu in unique_inc_pcus:
    df_selection = groep4DF_kz_pcu_df[groep4DF_kz_pcu_df['pcu'] == unique_pcu]

    # How often the unique_pcu is written
    total_length = len(df_selection)

    # How often the unique_pcu is written correct
    correct_count = len(df_selection[df_selection['correct'] == True])

    # How often the unique_pcu is written incorrect
    incorrect_count = len(df_selection[df_selection['correct'] == False])

    # In which prompts the PCU occurs
    prompt_occurence = ", ".join(list(set(list(df_selection['prompt']))))

    # List of PCUs with which target PCU is substituted
    unique_orig_graphs = df_selection['orig_pcu'].value_counts().index

    orig_pcu_string_list = []
    for unique_orig_graph in unique_orig_graphs:
        orig_pcu_string_list.append(unique_orig_graph + " (" + str(round((len(df_selection[df_selection['orig_pcu'] == unique_orig_graph])/total_length)*100, 2)) + ");")

    outputMatrix.append([unique_pcu, prompt_occurence, total_length, incorrect_count, round((incorrect_count/total_length)*100,2), len(orig_pcu_string_list)," ".join(orig_pcu_string_list)])

errorTableFile = os.path.join(lrecDataDir, '6-error-table-basiscript-g4p1.tsv')
errorTable = pd.DataFrame(outputMatrix, columns = ['pcu', 'prompts', 'count', 'inc_abs', 'inc_rel', 'count', 'substitutedby'])
errorTable["latex_seperator"] = ["\\\\"] * len(errorTable)
errorTable.to_csv(errorTableFile, sep='&', index=False)

print("See output table: ", errorTableFile)

OLD

# List all klankzuivere Target words
klankzuiverTargetDF = groep4DF[groep4DF['word_klankzuiver']]

# Klankzuivere woorden in dictee
list(set([x.replace("*", "") for x in list(klankzuiverTargetDF['target'])]))

# #Compute for each target&original word combo the number of correctly read klankzuivere PCUs

# def klankzuiverePCUreadCorrect(row):
#     target_phones = row['target_phones']
#     target_pcus = row['pcus_target_graph']
#     original_pcus = row['pcus_orig_graph']
#     klankzuiverheid_pcus = row['klankzuiver']

#     return sum([target_pcus[idx] == original_pcus[idx] and klankzuiverheid_pcus[idx] == True for idx, phone in enumerate(target_pcus)])

# klankzuiverTargetDF['klankzuiver_pcus_read_correct'] = klankzuiverTargetDF.apply(klankzuiverePCUreadCorrect, axis = 1)

# Wat voor soort fouten zijn er gemaakt?
def klankzuiverePCUreadIncorrect(row):
    target_phones = row['target_phones']
    target_pcus = row['pcus_target_graph']
    original_pcus = row['pcus_orig_graph']
    klankzuiverheid_pcus = row['klankzuiver']

    if len(target_phones) == len(target_pcus) and len(target_phones) == len(original_pcus):
        return [target_pcus[idx] + "_" + target_phones[idx] + "_" + original_pcus[idx] + "_" + "".join(target_pcus) + "_" + "".join(original_pcus) for idx, phone in enumerate(target_pcus) if (target_pcus[idx] != original_pcus[idx]) and (klankzuiverheid_pcus[idx] == True)]

    else:
        print(row["original"], target_phones, target_pcus, original_pcus, klankzuiverheid_pcus)

klankzuiverTargetDF.loc[:,'klankzuiver_pcus_read_incorrect'] = klankzuiverTargetDF.apply(klankzuiverePCUreadIncorrect, axis = 1)

list_incorrect_pcus_unflatten = [x for x in list(klankzuiverTargetDF['klankzuiver_pcus_read_incorrect']) if len(x) > 0]
print("nr of words with incorrect PCU:", len(list_incorrect_pcus_unflatten))

list_incorrect_pcus = [error.split('_') for error_item in list_incorrect_pcus_unflatten for error in error_item]
print("nr of incorrect PCUS: ", len(list_incorrect_pcus))

incorrectPCUdf = pd.DataFrame(list_incorrect_pcus, columns=['target_pcus', 'target_phones', 'orig_pcus', 'target_word', 'original_word'])

incorrectPCUdf[incorrectPCUdf['target_pcus']=='ch']

target_pcus_written_incorrect = list(incorrectPCUdf['target_pcus'].value_counts().index)

for target_pcu in target_pcus_written_incorrect:
    saved_list = []
    for pcu_list in klankzuiverTargetDF['target_pcus']:
        if target_pcu in pcu_list and "".join(pcu_list) not in saved_list:
            saved_list.append("".join(pcu_list))
    print(target_pcu, saved_list)


target_pcus_written_incorrect = list(incorrectPCUdf['target_pcus'].value_counts().index)

for target_pcu in target_pcus_written_incorrect:
    values = (incorrectPCUdf[incorrectPCUdf['target_pcus'] == target_pcu]['orig_pcus']).value_counts()
    print(target_pcu, len(list(values)))
