==> 3-dart-metadata-description.ipynb <==
import pandas as pd

# Read the data selection of 672 DART words and extract the filenames of the audio recordings
dart_word_selection_672 = '/vol/tensusers5/wharmsen/astla-data/lrec-data/4-pcu-segmentations-dart-preposttest.csv'
dart_word_selection_672_df = pd.read_csv(dart_word_selection_672, sep=';')
dart_word_selection_672_filenames = list(set([prompt_id.split('_')[0] + '.mp3' for prompt_id in dart_word_selection_672_df['filename']]))
print("# Unique DART filenames: ", len(dart_word_selection_672_filenames))
print(dart_word_selection_672_filenames)

# Read the audio file metadata
def getTestName(row):
    kern = row['Kern']
    exerciseID = row['ExerciseID']
    return kern+str(exerciseID)

dart_audio_metadata = '/vol/tensusers5/wharmsen/astla-data/dart-preposttest/specom-data/audio-metadata.csv'
dart_audio_metadata_df = pd.read_csv(dart_audio_metadata)
dart_audio_metadata_df['Testname'] = dart_audio_metadata_df.apply(getTestName, axis=1)
dart_audio_metadata_df

# Print for each audio recording used in the analysis which test the child read.
dart_audio_metadata_df_selection = dart_audio_metadata_df[dart_audio_metadata_df['Audio'].isin(dart_word_selection_672_filenames)]
print("Length of metadata selection:", len(dart_audio_metadata_df_selection))
dart_audio_metadata_df_selection.value_counts(['Kern', 'ExerciseID'])

# Add the name of the test the child had to read to the 672 words data selection
def addPreposttestName(row):
    prompt_id = row['filename']
    filename = prompt_id.split('_')[0] + '.mp3'

    # Find this file in audio metadata
    return dart_audio_metadata_df[dart_audio_metadata_df['Audio']==filename].reset_index().loc[0]['Testname']

# testname column
dart_word_selection_672_df['preposttest'] = dart_word_selection_672_df.apply(addPreposttestName, axis=1)
dart_word_selection_672_df.head()

# How many sound pure words does each test have?
for test in ['pretest0', 'pretest1', 'pretest2', 'posttest0', 'posttest1', 'posttest2']:
    print(test, len(list(set(dart_word_selection_672_df[(dart_word_selection_672_df['preposttest'] == test) & (dart_word_selection_672_df['word_klankzuiver'] == True)]['prompt']))))

dart_word_selection_672_df[dart_word_selection_672_df['word_klankzuiver'] == True]
